backend/requirements.txt
# Framework web - Versions sécurisées
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Base de données - Versions sécurisées
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Variables d'environnement
python-dotenv==1.0.1

# Authentification et sécurité - Versions sécurisées
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Validation des données - Versions sécurisées
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Utilitaires - Versions sécurisées
requests==2.32.3
httpx==0.28.1
aiohttp==3.11.11

# Sécurité supplémentaire
cryptography==44.0.0
bcrypt==4.2.1

# Monitoring et logging
structlog==24.4.0
sentry-sdk[fastapi]==2.19.2

# Rate limiting
slowapi==0.1.9

# Développement - Versions sécurisées
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0
bandit==1.7.10

# Production
gunicorn==23.0.0

# Monitoring et métriques
prometheus-client==0.20.0
psutil==5.9.8

# Cache et sessions
redis==5.0.7

# Monitoring d'erreurs
sentry-sdk[fastapi]==2.19.2

# Backup et stockage
boto3==1.35.84
botocore==1.35.84

# Logging structuré
structlog==24.4.0

# Tests et qualité
pytest-cov==5.0.0
coverage==7.6.4
